pyproject.toml
[build-system]
requires = [
    "setuptools>=40.8.0",
    "wheel"
]

[project]
name = "openthaigpt_model_test"
version = "0.1.0"
description = "OpenThaiGPT AI engineer exam."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "torch",
    "datasets",
    "transformers",
    "hydra-core",
    "pytorch-lightning",
]
